% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_boxp.R
\name{bdm.boxp}
\alias{bdm.boxp}
\title{Clustering statistics box-plot.}
\usage{
bdm.boxp(data, bdm, byVars = F, merged = T, clusters = NULL, layer = 1)
}
\arguments{
\item{data}{A matrix of data to be plotted (either the input data matrix or any covariate matrix as well).}

\item{bdm}{A \var{bdm} data mapping instance.}

\item{byVars}{Default value is \var{byVars = FALSE}, i.e. box-plots are grouped by cluster. If \var{byVars = TRUE} box-plots are grouped by input feature.}

\item{merged}{Default value is \var{merged = TRUE}, i.e. the box-plot depicts the clusters after merging (if merging has been performed). If \var{merged = FALSE} or no merging has been performed the box-plot depicts the top-level clustering.}

\item{clusters}{A vector with a subset of cluster ids. Default value is \var{clusters=NULL}, i.e. plot all clusters, (with a maximum of 25).}

\item{layer}{The ptSNE output layer. Default value is \var{layer = 1}.}
}
\value{
None.
}
\description{
Clustering statistics box-plot.
}
\details{
If the number of clusters is large, only the first 25 clusters will be plotted. Note that the WTT algorithm numbers the clusters based on density value at the peak cell of the cluster. Thus, the numbering of the clusters is highly correlated with their relevance in terms of partial density.
}
\examples{

bdm.example()
bdm.boxp(ex$data, ex$map)
bdm.boxp(ex$data, ex$map, byVars = TRUE)
}
