% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_dMap.R
\name{bdm.dMap.plot}
\alias{bdm.dMap.plot}
\title{Class density map plot.}
\usage{
bdm.dMap.plot(
  bdm,
  classes = NULL,
  join = FALSE,
  class.pltt = NULL,
  pakde.pltt = NULL,
  pakde.lvls = 16,
  wtt.lwd = 1,
  plot.peaks = T,
  labels.cex = 1,
  layer = 1
)
}
\arguments{
\item{bdm}{A \var{bdm} data mapping instance.}

\item{classes}{A vector with a subset of class names or covariate values. Default value is \var{classes=NULL}, i.e. all classes are plotted.}

\item{join}{Default value is \var{join = FALSE}, i.e. the class mapping is based on the class conditional distributions. If \var{join = TRUE}, class mapping is based on the overall join distribution.}

\item{class.pltt}{A colour palette to show class labels in the hard mapping. Default value is \var{class.pltt = NULL}, i.e. a default palette is used.}

\item{pakde.pltt}{A palette of colours to indicate the levels of the class density maps. The length of the colour palette should be at least the number of levels specified in \var{pakde.lvls}. Default value is \var{pakde.pltt = NULL}, i.e. a default palette is used.}

\item{pakde.lvls}{The number of levels of the heat-map when plotting class density maps. Default value is \var{pakde.lvls = 16}.}

\item{wtt.lwd}{The width of the watertrack lines (as set in \var{par()}). Default value is \var{wtt.lwd = 1.0}.}

\item{plot.peaks}{Default value is \var{plot.peaks = TRUE}, i.e. the density peak of each cluster is depicted.}

\item{labels.cex}{The size of the clusters labels (as set in \var{par()}). Default value is \var{labels.cex = 1.0}. If \var{labels.cex = 0.0} the labels of the clusters are not depicted.}

\item{layer}{The ptSNE output layer. Default value is \var{layer = 1}.}
}
\value{
None.
}
\description{
Class density map plot.
}
\details{
\var{bdm.dMap.plot()} yields a multi-plot layout where the first plot shows the dominating value of the covariate (or dominating class) in each cell, and the rest of the plots show the density map of each covariate value (or class).

The join distribution \eqn{P(V=v_{i},C=c_{j})} is prone to the bias in the marginal distribution of the covariate. Therefore, the join distribution \eqn{P(V=v_{i},C=c_{j})} is transformed, by default, into a conditional distribution \eqn{P(c_{j}|V=v_{i})} (where the \eqn{c_{j}} are the grid cells of the embedding and V is the covariate (or class)). Thus, the first plot shows a hard classification of grid-cells, (cells are coloured based on the dominating value of the covariate (or dominating class), \emph{i.e}. the \eqn{v_{i}} for which \eqn{P(c_{j}|V=v_{i})} is maximum), and the rest of the plots show the conditional distributions \eqn{P(C=c_{j}|V=v_{i})}. This makes the plots of the different classes not directly comparable but the dominant areas of each class can be more easily identified.

However, the same plots can be depicted based on the join distribution by setting \var{join = TRUE}. This makes sense when the bias in the covariate values (or classes) is not significant. In this case the hard clustering shows the real dominance of each covariate value (or class) over the embedding area and the density maps are comparable one to each other (although, individually, they are not real density functions as they do not add up to one).

The multi-plot layout can be limited to a subset of the values of the covariate (or subset of classes) specified in parameter \var{classes}.
}
\examples{

# --- load example dataset
bdm.example()
bdm.dMap.plot(ex$map)
}
