% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_main.R
\name{bdm.fName}
\alias{bdm.fName}
\title{Default \var{bdm} file name}
\usage{
bdm.fName(bdm)
}
\arguments{
\item{bdm}{A \var{bdm} instance as generated by \code{bdm.init()}.}
}
\value{
A \var{*.RData} file name based on \var{bdm$dSet} and main \var{bdm} parameters.
}
\description{
Generates a default file name. The default file name is intended for functions \code{bdm.save()} and \code{bdm.scp()} to ease the task of working/organizing multiple runs on the same dataset.
}
\details{
The file name is generated based on \code{bdm$dSet} and perplexity. If \code{!is.null(bdm$wtt)} the number of clusters is also included.
}
\examples{

bdm.example()
str(exMap$dSet)
str(exMap$ppx)
bdm.fName(exMap)
}
