% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_hlC.R
\name{bdm.hlCorr}
\alias{bdm.hlCorr}
\title{HD/LD correlation.}
\usage{
bdm.hlCorr(data, bdm, zSampleSize = 1000, threads = 4, mpi.cl = NULL)
}
\arguments{
\item{data}{Input data (a matrix, a big.matrix or a .csv file name).}

\item{bdm}{A \var{bdm} instance as generated by \code{bdm.ptsne()}.}

\item{zSampleSize}{Number of data points to check by thread. (Default value is \code{zSampleSize=1000}).}

\item{threads}{The number of parallel threads (according to data size and hardware resources, \code{i.e.} number of cores and available memory. Default value is \code{threads = 4}).}

\item{mpi.cl}{An MPI (inter-node parallelization) cluster as generated by \code{bdm.mpi.start()}. (By default \code{mpi.cl = NULL} a 'SOCK' (intra-node parallelization) cluster is generated).}
}
\value{
A copy of the input \var{bdm} instance with new element \var{bdm$knP}.
}
\description{
Pair-wise distance correlation between HD and LD neighborhoods.
}
\examples{

# --- load example dataset
\dontrun{
bdm.example()
m <- bdm.hlCorr(exData[, 1:4], ex$map, threads = 4)
}

}
