% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_knp.R
\name{bdm.knp}
\alias{bdm.knp}
\title{k-ary Neighborhood Preservation}
\usage{
bdm.knp(data, bdm, k.max = NULL, sampling = 0.9, mpi.cl = NULL, threads = 4)
}
\arguments{
\item{data}{Input data (a matrix, a big.matrix or a .csv file name).}

\item{bdm}{A \var{bdm} data mapping instance.}

\item{k.max}{Maximum neighborhood size to check. Default value is \var{k.max = NULL}, i.e. neighborhood sizes are checked up to \var{n-1}.}

\item{sampling}{Fraction of data points to check for each neighborhood size. Default value is \var{sampling = 0.9}.}

\item{mpi.cl}{An MPI (inter-node parallelization) cluster as returned by \var{bdm.mpi.start()}. Default value is \var{mpi.cl = NULL}, i.e. a 'SOCK' (intra-node parallelization) cluster is generated.}

\item{threads}{The number of parallel threads (according to data size and hardware resources, i.e. number of cores and available memory). Default value is \var{threads = 4}.}
}
\value{
A \var{bdm} data mapping instance.
#'
}
\description{
A measure of matching between HD and LD neighborhoods ('Multi-scale similarities in stochastic neighbour embedding: Reducing dimensionality while preserving both local and global structure', Lee et. al 2015).
}
\examples{

# --- load example dataset
bdm.example()
\dontrun{
# --- compte the kNP
m <- bdm.knp(ex$data, ex$map, threads = 4)
# --- plot the kNP
bdm.knp.plot(m)
}

}
