% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_env.R
\name{bdm.local}
\alias{bdm.local}
\title{Set/get default local machine name or IP address}
\usage{
bdm.local(dest = NULL)
}
\arguments{
\item{dest}{Name or IP address of the local machine.}
}
\value{
The current value of \var{local}
}
\description{
Set/get default local machine name or IP address
}
\examples{

# --- set default value of \\var{local}
bdm.local('xxx.255.0.0')
bdm.local('mymachine.mydomain.cat')
}
