% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_merge.R
\name{bdm.merge.s2nr}
\alias{bdm.merge.s2nr}
\title{Merging of clusters based on signal-to-noise-ratio.}
\usage{
bdm.merge.s2nr(
  data,
  bdm,
  k = 10,
  plot.merge = T,
  ret.merge = F,
  info = T,
  layer = 1,
  ...
)
}
\arguments{
\item{data}{Input data (a matrix, a big.matrix or a .csv file name).}

\item{bdm}{A \var{bdm} data mapping instance.}

\item{k}{The number of desired clusters. The clustering will be recursively merged until reaching this number of clusters. Default value is \var{k = 10}. By setting \var{k < 0} we specify the number of clusters we want to merge (as opposed to the number of final clusters).}

\item{plot.merge}{Default value is \var{plot.merge = TRUE}, i.e. the merged clustering is plotted.}

\item{ret.merge}{Default value is \var{ret.merge = FALSE}. If \var{ret.merge = TRUE}, the function returns a \var{bdm} data mapping instance with the merged clustering.}

\item{info}{Default value is \var{info = FALSE}. If TRUE, all merging steps are shown.}

\item{layer}{The ptSNE output layer. Default value is \var{layer = 1}.}

\item{...}{If \var{plot.merge} is TRUE, you can pass plotting parameters to \var{bdm.wtt.plot()}.}
}
\value{
None if \var{ret.merge = FALSE}, else, a \var{bdm} data mapping instance.
}
\description{
Performs a recursive merging of clusters based on minimum loss of signal-to-noise-ratio (S2NR) until reaching the desired number of clusters. The S2NR is the explained/unexplained variance ratio measured in the high dimensional space based on the given low dimensional clustering.
}
\details{
See details in \var{bdm.optk.s2nr()}.
}
\examples{

# --- load mapped dataset
bdm.example()
\dontrun{
# --- merge and plot (no save)
bdm.merge.s2nr(ex$data, ex$map, k = 12)
# --- merge and return merging (do no plot)
m <- bdm.merge.s2nr(ex$data, ex$map, k = 12, ret.merge = T, plot.merge = F)
# --- plot merging afterwards (use bdm.wtt.plot() function)
bdm.wtt.plot(m)
}
}
