% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_clst.R
\name{bdm.mpi.start}
\alias{bdm.mpi.start}
\title{Initialize parallel computing environment.}
\usage{
bdm.mpi.start(threads)
}
\arguments{
\item{threads}{The number of parallel threads (in principle only limited by hardware resources, \code{i.e.} number of cores and available memory)}
}
\value{
cl A cluster instance (as created by the snow::makeCluster() function).
}
\description{
Initialize parallel computing environment.
}
