% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_clst.R
\name{bdm.mpi.stop}
\alias{bdm.mpi.stop}
\title{Stops MPI parallel computing environment.}
\usage{
bdm.mpi.stop(cl)
}
\arguments{
\item{cl}{A cluster instance (as created by the bdm.mpi.start() function).}
}
\description{
Stops MPI parallel computing environment.
}
