% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_env.R
\name{bdm.mybdm}
\alias{bdm.mybdm}
\title{Set/get default path for \var{mybdm}}
\usage{
bdm.mybdm(path = NULL)
}
\arguments{
\item{path}{Path to \var{mybdm}.}
}
\value{
The current path value to \var{mybdm}
}
\description{
Set/get default path for \var{mybdm}
}
\examples{

# --- set default path for \\var{mybdm}
bdm.mybdm('~/mybdm')
}
