% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_main.R
\name{bdm.pakde}
\alias{bdm.pakde}
\title{Perplexity-adaptive kernel density estimation}
\usage{
bdm.pakde(
  bdm,
  ppx = 100,
  g = 200,
  g.exp = 3,
  mpi.cl = NULL,
  threads = 2,
  layer = 1
)
}
\arguments{
\item{bdm}{A \var{bdm} data mapping instance.}

\item{ppx}{The value of perplexity to compute similarities in the low-dimensional embedding. Default value is \var{ppx = 100}.}

\item{g}{The resolution of the density space grid (\eqn{g*g} cells). Default value is \var{g = 200}.}

\item{g.exp}{A numeric factor to avoid border effects. The grid limits will be expanded so as to enclose the density of the kernel of the most extreme embedded datapoints up to \var{g.exp} times \eqn{\sigma}. Default value is \var{g.exp = 3}, \var{i.e.} the grid limits are expanded so as to enclose the 0.9986 of the probability mass of the most extreme kernels.}

\item{mpi.cl}{An MPI (inter-node parallelization) cluster as returned by \var{bdm.mpi.start()}. Default value is \var{mpi.cl = NULL}, i.e. a 'SOCK' (intra-node parallelization) cluster is automatically generated.}

\item{threads}{Number of parallel threads (according to data size and hardware resources, i.e. number of cores and available memory). Default value is \var{threads = 4}.}

\item{layer}{The ptSNE output layer. Default value is \var{layer = 1}.}
}
\value{
A copy of the input \var{bdm} instance with new element \var{bdm$pakde} (paKDE output). \var{bdm$pakde[[layer]]$layer = 'NC'} stands for not computed layers.
}
\description{
Starts the paKDE algorithm (second step of the mapping protocol).
}
\details{
When computing the \var{paKDE} the embedding area is discretized as a grid of size \var{g*g} cells. In order to avoid border effects, the limits of the grid are expanded by default so as to enclose at least the 0.9986 of the cumulative distribution function (\eqn{3 \sigma}) of the kernels of the most extreme mapped points in each direction.

The presence of outliers in the embedding can lead to undesired expansion of the grid limits. We can overcome this using lower values of \var{g.exp}. By setting \var{g.exp = 0} the grid limits will be equal to the range of the embedding.

The values \var{g.exp = c(1, 2, 3, 4, 5, 6)} enclose cdf values of \var{0.8413, 0.9772, 0.9986, 0.99996, 0.99999, 1.0} respectively.
}
\examples{

# --- load mapped dataset
bdm.example()
# --- run paKDE
\dontrun{
m <- bdm.pakde(ex$map, ppx = 200, g = 200, g.exp = 3, threads = 4)
# --- plot paKDE output
bdm.pakde.plot(m)
}
}
