% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_qMap.R
\name{bdm.pMap}
\alias{bdm.pMap}
\title{Precision map (quantile map of betas)}
\usage{
bdm.pMap(
  bdm,
  pMap.levels = 8,
  pMap.cex = 0.1,
  pMap.bg = "#000000",
  colorbar = T
)
}
\arguments{
\item{bdm}{A \var{bdm} instance as generated by \code{bdm.init()}.}

\item{pMap.levels}{The number of levels of the quantile-map (8 by default).}

\item{pMap.cex}{The size of the data-points (as in \code{par()}). Default value is \code{ptsne.cex = 0.1}.}

\item{pMap.bg}{The background colour of the qMap plot. Default value is \code{ptsne.bg = #FFFFFF} (white).}

\item{colorbar}{A logical value (TRUE by default). FALSE hides the side colorbar.}
}
\value{
None.
}
\description{
Precision map (quantile map of betas)
}
\examples{

bdm.example()
bdm.pMap(ex$map)
}
