% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_plot.R
\name{bdm.ptsne.plot}
\alias{bdm.ptsne.plot}
\title{Plot ptSNE (low-dimensional embedding)}
\usage{
bdm.ptsne.plot(
  bdm,
  ptsne.cex = 0.5,
  ptsne.bg = "#FFFFFF",
  class.lbls = NULL,
  class.pltt = NULL,
  layer = 1
)
}
\arguments{
\item{bdm}{A \var{bdm} data mapping instance, or a list of them to make a comparative plot.}

\item{ptsne.cex}{The size of the mapped data-points in the ptSNE plot. Default value is \var{ptsne.cex = 0.5}.}

\item{ptsne.bg}{The background colour of the ptSNE plot. Default value is \var{ptsne.bg = #FFFFFF} (white).}

\item{class.lbls}{A vector of class labels to show in the ptSNE plot. Default value is \var{class.lbls = NULL}. If \var{is.null(class.lbls)} and \var{!is.null(bdm$lbls)} the later will be used. If \var{!is.null(bdm$lbls)} and \var{!is.null(bdm$wtt)} (i.e. there is a clustering) cluster labels are used by default.}

\item{class.pltt}{A colour palette to show class labels in the ptSNE plot. If \var{!is.null(bdm$wtt)} cluster labels are used by default, else if \var{!is.null(bdm$lbls)} are used by default. If \var{ptsne.pltt = NULL} (default value) the default palette is used.}

\item{layer}{The ptSNE output layer. Default value is \var{layer = 1}.}
}
\value{
None.
}
\description{
Plot ptSNE (low-dimensional embedding)
}
\examples{

bdm.example()
bdm.ptsne.plot(ex$map)
}
