% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_main.R
\name{bdm.restart}
\alias{bdm.restart}
\title{Restart pt-SNE}
\usage{
bdm.restart(
  dSet.data,
  bdm,
  epochs = NULL,
  iters = NULL,
  mpi.cl = NULL,
  threads = NULL,
  layers = NULL,
  info = 0
)
}
\arguments{
\item{dSet.data}{Input data (a matrix, a big.matrix or a .csv file name).}

\item{bdm}{A \var{bdm} instance as generated by \code{bdm.init()}.}

\item{epochs}{Number of epochs to run. Default value \code{epochs = NULL} runs \code{4 *log(n)} epochs.}

\item{iters}{Number of iters/epoch. Default value \code{iters = NULL} runs \code{4 *log(thread_size)} iters/epoch.}

\item{mpi.cl}{MPI (inter-node parallelization) cluster as generated by \code{bdm.mpi.start()}. (By default \code{mpi.cl = NULL} a 'SOCK' (intra-node parallelization) cluster is generated).}

\item{threads}{Number of parallel threads (according to data size and hardware resources, \code{i.e.} number of cores and available memory. Default value is \code{threads = 4}).}

\item{layers}{Number of layers (\code{minimum} 2, \code{maximum} the number of threads).}

\item{info}{Output information: 1 yields inter-round results, 0 disables intermediate results. Default value is 0.}
}
\value{
A copy of the input \var{bdm} instance with new element \var{bdm$ptsne} (t-SNE output).
}
\description{
Restarts the ptSNE algorithm (runs more epochs).
}
\examples{

# --- load example dataset
bdm.example()
# --- perform ptSNE
\dontrun{
m <- bdm.restart(ex$data, ex$map, epochs = 50)
}
}
