% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_main.R
\name{bdm.save}
\alias{bdm.save}
\title{Save \var{bdm} instance}
\usage{
bdm.save(...)
}
\arguments{
\item{...}{A \var{bdm} instance as generated by \code{bdm.init()}.}
}
\value{
None
}
\description{
Saves a \var{bdm} instance with default path/file names, as given by \code{bdm.mybdm()/bdm.fName(bdm)}. Default file name is generated based on \code{bdm$dSet} and perplexity.  The purpose of functions \code{bdm.save()} and \code{bdm.scp()} used with \code{bdm.fName()} is to ease the task of working/organizing multiple runs on the same dataset.
}
\examples{

bdm.example()
# --- save it
\dontrun{
mybdm <- bdm.init(exData[, 1:4], dSet.name = 'myEx', ppx = 50, dSet.labels = exData[, 5])
str(mybdm)
bdm.save(mybdm)
}
}
