% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_plot.R
\name{bdm.wtt.plot}
\alias{bdm.wtt.plot}
\title{Plot WTT (clustering)}
\usage{
bdm.wtt.plot(
  bdm,
  pakde.pltt = NULL,
  pakde.lvls = 16,
  wtt.lwd = 1,
  plot.peaks = T,
  labels.cex = 1,
  layer = 1
)
}
\arguments{
\item{bdm}{A \var{bdm} data mapping instance, or a list of them to make a comparative plot.}

\item{pakde.pltt}{A colour palette to show levels in the paKDE plot. By default (\var{pakde.pltt = NULL}) the default palette is used.}

\item{pakde.lvls}{The number of levels of the density heat-map (16 by default).}

\item{wtt.lwd}{The width of the watertrack lines (as set in \var{par()}).}

\item{plot.peaks}{Default value is \var{plot.peaks = TRUE}. If TRUE the density peaks are indicated.}

\item{labels.cex}{If \var{plot.peaks} is TRUE, the size of the peaks labels (as set in \var{par()}). Default value is \var{labels.cex = 1.0}.}

\item{layer}{The ptSNE output layer. Default value is \var{layer = 1}.}
}
\value{
None.
}
\description{
Plot WTT (clustering)
}
\examples{

bdm.example()
bdm.wtt.plot(ex$map)
}
