% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_main.R
\name{bdm.wtt}
\alias{bdm.wtt}
\title{Watertrack transform (WTT)}
\usage{
bdm.wtt(bdm, layer = 1)
}
\arguments{
\item{bdm}{A \var{bdm} instance as generated by \code{bdm.init()}.}

\item{layer}{The number of the t-SNE layer. Default value is \code{layer = 1}.}
}
\value{
A copy of the input \var{bdm} instance with \var{bdm$wtt} (WTT output).
}
\description{
Starts the WTT algorithm (third setp of the mapping protocol).
}
\details{
This function requires the up-stream step \code{bdm.pakde()}.
}
\examples{

# --- load mapped dataset
bdm.example()
# --- perform WTT
m <- bdm.wtt(ex$map)
# --- plot WTT output
bdm.wtt.plot(m)
}
