% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_boxp.R
\name{bdm.boxp}
\alias{bdm.boxp}
\title{Clustering statistics box-plot.}
\usage{
bdm.boxp(bdm, byVars = F, layer = 1)
}
\arguments{
\item{bdm}{A \var{bdm} instance as generated by \code{bdm.init()}.}

\item{byVars}{A logical value. By default (\code{byVars = FALSE}) box-plots are grouped by cluster. With \code{byVars = TRUE} box-plots are grouped by input feature.}

\item{layer}{The number of a layer (1 by default).}
}
\value{
None.
}
\description{
Clustering statistics box-plot.
}
\details{
If the number of clusters is large, only the first 25 clusters will be plotted. Note that the WTT algorithm numbers the clusters based on density value at the peak cell of the cluster. Thus, the numbering of the clusters is highly correlated with their relevance in terms of partial density. Therefore, in case of more than 25 clusters, the most relevant should always be included in the plot.
}
\examples{

bdm.example()
bdm.boxp(exMap)
bdm.boxp(exMap, byVars = TRUE)
}
