% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_plot.R
\name{bdm.cost}
\alias{bdm.cost}
\title{ptSNE cost & size plot.}
\usage{
bdm.cost(bdm, offset = 0)
}
\arguments{
\item{bdm}{A \var{bdm} instance as generated by \code{bdm.init()} or a list of them to make a comparative plot.}

\item{offset}{X-axis offset in number of epochs (0 by default).}
}
\value{
None.
}
\description{
ptSNE cost & size plot.
}
\examples{

bdm.example()
bdm.cost(exMap)
}
