% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_plot.R
\name{bdm.cost}
\alias{bdm.cost}
\title{ptSNE cost & size plot.}
\usage{
bdm.cost(bdm, x.lim = NULL)
}
\arguments{
\item{bdm}{A \var{bdm} instance as generated by \code{bdm.init()} or a list of them to make a comparative plot.}

\item{x.lim}{A vector with upper and lower bounds to limit the number of epochs in the x-axis (NULL by default).}
}
\value{
None.
}
\description{
ptSNE cost & size plot.
}
\examples{

bdm.example()
bdm.cost(ex$map)
}
