% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_plot.R
\name{bdm.cost}
\alias{bdm.cost}
\title{ptSNE cost & size plot.}
\usage{
bdm.cost(bdm, x.lim = NULL)
}
\arguments{
\item{bdm}{A \var{bdm} data mapping instance, or a list of them to make a comparative plot.}

\item{x.lim}{X-axis (epochs) limits. Default value is \var{x.lim = NULL}.}
}
\value{
None.
}
\description{
ptSNE cost & size plot.
}
\examples{

bdm.example()
bdm.cost(ex$map)
}
