% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_main.R
\name{bdm.example}
\alias{bdm.example}
\title{Example dataset}
\usage{
bdm.example()
}
\value{
An example dataset named \code{ex}
}
\description{
Loads a mapping example.
}
\details{
Object \code{ex} is a list with elements: \var{ex$data} a matrix with raw data; \var{ex$labels} a vector of datapoint labels; \var{ex$map} is a \code{bdm} instance: a list with the results of the mapping protocol.

This example is based on a small synthetic dataset with \code{n = 5000} observations drawn from a 4-variate Gaussian Mixture Model (GMM) with 16 Gaussian components with random means and variances.
}
\examples{

# --- load example dataset
bdm.example()
str(ex)
}
