% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_hlC.R
\name{bdm.hlCorr}
\alias{bdm.hlCorr}
\title{HD/LD correlation.}
\usage{
bdm.hlCorr(data, bdm, zSampleSize = 1000, mpi.cl = NULL, threads = 4)
}
\arguments{
\item{data}{Input data (a matrix, a big.matrix or a .csv file name).}

\item{bdm}{A \var{bdm} data mapping instance.}

\item{zSampleSize}{Number of data points to check by thread. Default value is \var{zSampleSize = 1000}.}

\item{mpi.cl}{An MPI (inter-node parallelization) cluster as returned by \var{bdm.mpi.start()}. Default value is \var{mpi.cl = NULL}, i.e. a 'SOCK' (intra-node parallelization) cluster is generated.}

\item{threads}{The number of parallel threads (according to data size and hardware resources, i.e. number of cores and available memory). Default value is \var{threads = 4}.}
}
\value{
A \var{bdm} data mapping instance.
}
\description{
Pair-wise distance correlation between HD and LD neighborhoods.
}
\examples{

# --- load example dataset
\dontrun{
bdm.example()
m <- bdm.hlCorr(ex$data, ex$map, threads = 4)
}

}
