% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_main.R
\name{bdm.init}
\alias{bdm.init}
\title{Embedding initialization.}
\usage{
bdm.init(
  data,
  is.distance = F,
  is.sparse = F,
  ppx = 100,
  mpi.cl = NULL,
  threads = 4,
  labels = NULL
)
}
\arguments{
\item{data}{A \var{data.frame} or \var{matrix} with raw input-data. The dataset must not have duplicated rows.}

\item{is.distance}{Default value is \var{is.distance = FALSE}. TRUE indicates that raw data is a distance matrix.}

\item{is.sparse}{Default value is \var{is.sparse = FALSE}. TRUE indicates that the raw data is a sparse matrix.}

\item{ppx}{The value of perplexity to compute similarities.}

\item{mpi.cl}{An MPI (inter-node parallelization) cluster as returned by \var{bdm.mpi.start()}. By default \var{mpi.cl = NULL}, i.e. a 'SOCK' (intra-node parallelization) cluster is used.}

\item{threads}{Number of parallel threads (according to data size and hardware resources, \var{i.e.} number of cores and available memory. Default value is \var{threads = 4}).}

\item{labels}{If available, a length \var{nrow(data)} vector of class labels. Label values are factorized as \var{as.numeric(as.factor(labels))}.}
}
\value{
A \var{bdm} data mapping instance. This \var{bdm} instance is the starting object of the mapping protocol: a list to which new elements are added at each step of the mapping protocol.
}
\description{
Computes the precision parameters for the given perplexity (i.e. the local bandwidths for the input affinity kernels) and returns them as a \var{bdm} data mapping instance. A \var{bdm} data mapping instance is the starting object of the mapping protocol, a list to which new elements are added at each step of the mapping protocol.
}
\examples{

# --- load example dataset
bdm.example()
\dontrun{
m <- bdm.init(ex$data, ppx = 250, labels = ex$labels)
}
}
