% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_main.R
\name{bdm.init}
\alias{bdm.init}
\title{Create \var{bdm} instance}
\usage{
bdm.init(
  dSet.data,
  dSet.name = NULL,
  is.distance = F,
  is.sparse = F,
  normalize = F,
  ppx = 100,
  xppx = 3,
  dSet.labels = NULL,
  threads = 4,
  mpi.cl = NULL
)
}
\arguments{
\item{dSet.data}{A \var{data.frame} or \var{matrix} with raw input-data. The dataset must not have duplicated rows.}

\item{dSet.name}{The name given to the input dataset (optional). This name will be used to automatically generate a name to save the output as an \var{.Rdata} file.}

\item{is.distance}{A logical value (FALSE by default). TRUE indicates that raw data is a distance matrix.}

\item{is.sparse}{A logical value (FALSE by default). TRUE indicates that the raw data is a sparse matrix.}

\item{normalize}{A logical value (TRUE by default). By default data is normalized before computing affinities. Set to FALSE to override this default behaviour (for instance if the input data is already normalized).}

\item{ppx}{The value of perplexity to compute similarities (default value is 100).}

\item{dSet.labels}{If available, labels can be included as a separate vector of length equal to \code{nrow(dSet.data)}. Label values are factorized as \code{as.numeric(as.factor(labels))}.}

\item{threads}{Number of parallel threads (according to data size and hardware resources, \code{i.e.} number of cores and available memory. Default value is \code{threads = 4}).}

\item{mpi.cl}{MPI (inter-node parallelization) cluster as generated by \code{bdm.mpi.start()}. (By default \code{mpi.cl = NULL} a 'SOCK' (intra-node parallelization) cluster is generated).}
}
\value{
A \var{bdm} instance. A \var{bdm} instance is a list with the supplied info and the precision parameters computed for the given perplexity. This list is the starting object to which new elements are added at each step of the mapping protocol.
}
\description{
Computes the precision parameters beta_i for the given perplexity (local bandwidths for the input affinity kernels) and returns them in a \var{bdm} instance. A \var{bdm} instance is a list with the supplied info (type of inpu data and perplexity) and the precision parameters that have been computed. This list is the starting object to which new elements are added at each step of the mapping protocol.
}
\examples{

# --- load example dataset
bdm.example()
# --- perform ptSNE
\dontrun{
myMap <- bdm.init(exData[, 1:4], dSet.name = 'ex', ppx = 250, labels = exData[, 5])
}
}
