% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_knp.R
\name{bdm.knp.plot}
\alias{bdm.knp.plot}
\title{k-ary Neighborhood Preservation plot}
\usage{
bdm.knp.plot(bdm)
}
\arguments{
\item{bdm}{A \var{bdm} data mapping instance, or a list of them to make a comparative plot.}
}
\description{
Log and linear plots of the k-ary Neighborhood Preservation
}
\examples{

# --- load example dataset
bdm.example()
\dontrun{
# --- compte the kNP
m <- bdm.knp(ex$data, ex$map, threads = 4)
# --- plot the kNP
bdm.knp.plot(m)
}

}
