% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_knp.R
\name{bdm.knp.plot}
\alias{bdm.knp.plot}
\title{k-ary Neighborhood Preservation plot}
\usage{
bdm.knp.plot(bdm)
}
\arguments{
\item{bdm}{A \var{bdm} instance as generated by \code{bdm.ptsne()} or a list of them.}
}
\description{
Log and linear plots of the k-ary Neighborhood Preservation
}
\examples{

# --- load example dataset
\dontrun{
bdm.example()
m <- bdm.knp(ex$data, ex$map, threads = 4)
bdm.knp.plot(m)
}

}
