% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_knp.R
\name{bdm.knp}
\alias{bdm.knp}
\title{k-ary Neighborhood Preservation}
\usage{
bdm.knp(data, bdm, k.max = NULL, sampling = 0.9, threads = 4, mpi.cl = NULL)
}
\arguments{
\item{data}{Input data (a matrix, a big.matrix or a .csv file name).}

\item{bdm}{A \var{bdm} instance as generated by \code{bdm.ptsne()}.}

\item{k.max}{Maximum neighborhood size to check. (By default \code{k.max=NULL} neighborhood sizes are checked up to n-1).}

\item{sampling}{Fraction of data points to check for each neighborhood size. (Default value is \code{sampling=0.9}).}

\item{threads}{The number of parallel threads (according to data size and hardware resources, \code{i.e.} number of cores and available memory. Default value is \code{threads = 4}).}

\item{mpi.cl}{An MPI (inter-node parallelization) cluster as generated by \code{bdm.mpi.start()}. (By default \code{mpi.cl = NULL} a 'SOCK' (intra-node parallelization) cluster is generated).}
}
\value{
A copy of the input \var{bdm} instance with new element \var{bdm$knP}.
#'
}
\description{
A measure of matching between HD and LD neighborhoods ('Multi-scale similarities in stochastic neighbour embedding: Reducing dimensionality while preserving both local and global structure', Lee et. al 2015).
}
\examples{

# --- load example dataset
\dontrun{
bdm.example()
m <- bdm.knp(exData[, 1:4], exMap, threads = 4)
}

}
