% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_main.R
\name{bdm.labels}
\alias{bdm.labels}
\title{Get data-point clustering labels.}
\usage{
bdm.labels(bdm, merged = T, layer = 1)
}
\arguments{
\item{bdm}{A \var{bdm} data mapping instance.}

\item{merged}{Default value is \var{merged = TRUE}. If \var{merged = TRUE} and the clustering has been merged, the labels are the ids of the clusters after merging. If \var{merged = FALSE} or the clustering has not been merged, the labels indicate the ids of to the top-level clustering.}

\item{layer}{The ptSNE output layer. Default value is \var{layer = 1}.}
}
\value{
A vector of data-point clustering labels.
}
\description{
Given that clusters are computed at grid-cell level, this function returns the clustering label for each data-point.
}
\examples{

bdm.example()
m.labels <- bdm.labels(ex$map)
}
