% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_main.R
\name{bdm.labels}
\alias{bdm.labels}
\title{Get data-point clustering labels.}
\usage{
bdm.labels(bdm, merged = T, layer = 1)
}
\arguments{
\item{bdm}{A \var{bdm} instance as generated by \code{bdm.init()}.}

\item{merged}{A logical value. If TRUE (default value) and the \code{!is.null(bdm$merge)}, labels are the ids of the clusters after merging. If \code{merged = FALSE} or \code{is.null(bdm$merge)} labels indicate the ids of to the top-level clustering.}

\item{layer}{The number of the t-SNE layer (1 by default).}
}
\value{
A vector of data-point clustering labels.
}
\description{
Given that clusters are computed at grid-cell level, this function returns the clustering label for each data-point.
}
\examples{

bdm.example()
exMap.labels <- bdm.labels(exMap)
}
