% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_merge.R
\name{bdm.merge.s2nr}
\alias{bdm.merge.s2nr}
\title{Merging of clusters based on signal-to-noise-ratio.}
\usage{
bdm.merge.s2nr(
  data,
  bdm,
  k = 10,
  plot.merge = T,
  ret.merge = F,
  info = T,
  layer = 1,
  ...
)
}
\arguments{
\item{data}{Input data (a matrix, a big.matrix or a .csv file name).}

\item{bdm}{A \var{bdm} instance as generated by \code{bdm.init()}.}

\item{k}{The number of desired clusters. The clustering will be recursively merged until reaching this number of clusters. Default value \code{k = 10}. By setting \code{k < 0} we specify the number of clusters that we want to merge.}

\item{plot.merge}{Default value is \code{plot.merge = TRUE}. If TRUE, the merged clustering is plotted.}

\item{ret.merge}{Default value is \code{ret.merge = FALSE}. If TRUE, the function returns a \var{bdm} instance with the merged clustering.}

\item{info}{Default value is \code{info = FALSE}. If TRUE, all merging steps are shown.}

\item{layer}{The \var{bdm$ptsne} layer to be used. Default value is \code{layer = 1}.}

\item{...}{If \var{plot.merge} is TRUE, you can pass plotting parameters to \code{bdm.wtt.plot()}.}
}
\value{
None if \code{ret.merge = FALSE}. Else, a \var{bdm} instance with the merged clustering (\var{bdm$merge}).
}
\description{
Performs a recursive merging of clusters based on minimum loss of signal-to-noise-ratio (S2NR) until reaching the desired number of clusters. The S2NR is the explained/unexplained variance ratio measured in the high dimensional space based on the given low dimensional clustering.
}
\details{
See details in \code{bdm.optk.s2nr()}.
}
\examples{

bdm.example()
# --- merge and plot (no save)
bdm.merge.s2nr(ex$data, ex$map, k = 12)
\dontrun{
# --- merge and return merging (do no plot)
m <- bdm.merge.s2nr(ex$data, ex$map, k = 12, ret.merge = T, plot.merge = F)
# --- plot merging afterwards (use bdm.wtt.plot() function)
bdm.wtt.plot(m)
}
}
