% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_merge.R
\name{bdm.optk.plot}
\alias{bdm.optk.plot}
\title{Plots the signal-to-nois-ratio as a function of the number of clusters.}
\usage{
bdm.optk.plot(bdm)
}
\arguments{
\item{bdm}{A \var{bdm} instance as generated by \code{bdm.init()}.}
}
\value{
None.
}
\description{
The function \code{bdm.optk.sn2r()} computes the S2NR that results from recursively merging clusters and, by deafult, makes a plot of these values. For large datasets this computation can take a while, so we can save this result by setting \code{ret.optk = TRUE}. If this result is saved, we can plot it again at any time using this funcion.
}
\examples{

bdm.example()
exMap <- bdm.optk.s2nr(exMap, ret.optk = TRUE)
bdm.optk.plot(exMap)
}
