% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_main.R
\name{bdm.ptsne}
\alias{bdm.ptsne}
\title{Parallelized t-SNE (ptSNE)}
\usage{
bdm.ptsne(
  dSet.data,
  bdm,
  theta = 0.5,
  Y.init = NULL,
  mpi.cl = NULL,
  threads = 4,
  layers = 2,
  info = 0
)
}
\arguments{
\item{dSet.data}{Input data (a matrix, a big.matrix or a .csv file name).}

\item{bdm}{A \var{bdm} instance as generated by \code{bdm.init()}.}

\item{theta}{Accuracy/speed trade-off factor, a value between 0.33 and 0.8. (Default value is \code{theta = 0.0}). If \code{theta < 0.33} the algorithm uses the exact computation of the gradient. The closer is this value to 1 the faster is the computation but the coarser is the approximation of the gradient.}

\item{Y.init}{A \code{n *2 *layers} matrix with initial mapping positions. (By default \code{Y.init=NULL} will use random initial positions).}

\item{mpi.cl}{MPI (inter-node parallelization) cluster as generated by \code{bdm.mpi.start()}. (By default \code{mpi.cl = NULL} a 'SOCK' (intra-node parallelization) cluster is generated).}

\item{threads}{Number of parallel threads (according to data size and hardware resources, \code{i.e.} number of cores and available memory. Default value is \code{threads = 4}).}

\item{layers}{Number of layers (\code{minimum} 2, \code{maximum} the number of threads).}

\item{info}{Output information: 1 yields inter-round results, 0 disables intermediate results. Default value is 0.}
}
\value{
A copy of the input \var{bdm} instance with new element \var{bdm$ptsne} (t-SNE output).
}
\description{
Starts the parallelized t-SNE algorithm (pt-SNE). This is the first step of the mapping protocol.
}
\examples{

# --- load example dataset
bdm.example()
# --- perform ptSNE
\dontrun{
m <- bdm.ptsne(ex$data, ex$map, threads = 10, layers = 2)
# --- plot the Cost function
bdm.cost(m)
# --- plot ptSNE output
bdm.ptsne.plot(m, class.lbls = ex$labels)
}
}
