% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_main.R
\name{bdm.ptsne}
\alias{bdm.ptsne}
\title{Parallelized t-SNE}
\usage{
bdm.ptsne(
  data,
  bdm,
  lRate = NULL,
  theta = 0,
  alpha = 0.5,
  Y.init = NULL,
  info = 0,
  threads = 3,
  mpi.cl = NULL,
  layers = 2,
  rounds = 1
)
}
\arguments{
\item{data}{Input data (a matrix, a big.matrix or a .csv file name).}

\item{bdm}{A \var{bdm} instance as generated by \code{bdm.init()}.}

\item{lRate}{Learning-rate. (By default \code{lRate=NULL} it is adjusted to \code{n *layers /threads /16}.}

\item{theta}{Accuracy/speed trade-off factor, a value between 0.33 and 0.8. (Default value is \code{theta = 0.0}). If \code{theta < 0.33} the algorithm uses the exact computation of the gradient. The closer is this value to 1 the faster is the computation but the coarser is the approximation of the gradient.}

\item{alpha}{Momentum factor (Default value is \code{alpha=0.5}).}

\item{Y.init}{A \code{n *2 *layers} matrix with initial mapping positions. (By default \code{Y.init=NULL} will use random initial positions).}

\item{info}{Output information: 1 yields inter-round results, 0 disables intermediate results. Default value is 0.}

\item{threads}{Number of parallel threads (according to data size and hardware resources, \code{i.e.} number of cores and available memory. Default value is \code{threads = 4}).}

\item{mpi.cl}{MPI (inter-node parallelization) cluster as generated by \code{bdm.mpi.start()}. (By default \code{mpi.cl = NULL} a 'SOCK' (intra-node parallelization) cluster is generated).}

\item{layers}{Number of layers (\code{minimum} 2, \code{maximum} the number of threads).}

\item{rounds}{Number of rounds (1 by default).}
}
\value{
A copy of the input \var{bdm} instance with new element \var{bdm$ptsne} (t-SNE output).
}
\description{
Starts the ptSNE algorithm (first step of the mapping protocol).
}
\examples{

# --- load example dataset
bdm.example()
# --- perform ptSNE
\dontrun{
exMap <- bdm.ptsne(exMap, threads = 10, layers = 2, rounds = 2, ppx = 200)
}
# --- plot the Cost function
bdm.cost(exMap)
# --- plot ptSNE output
bdm.ptsne.plot(exMap)
}
