% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_qMap.R
\name{bdm.qMap}
\alias{bdm.qMap}
\title{ptSNE quantile-maps}
\usage{
bdm.qMap(
  bdm,
  data,
  labels = NULL,
  subset = NULL,
  qMap.levels = 8,
  qMap.cex = 0.3,
  qMap.bg = "#FFFFFF",
  class.pltt = NULL,
  qtitle = NULL,
  cex.main = 1,
  colorbar = T,
  layer = 1
)
}
\arguments{
\item{bdm}{A \var{bdm} instance as generated by \code{bdm.init()}.}

\item{data}{A \code{matrix/data.frame} to be mapped.}

\item{labels}{A length \code{nrow(bdm$data)} vector of class labels to overlay onto the embedding. Label values are factorized as \var{as.numeric(as.factor(labels))}. Default value is \var{labels = NULL}.}

\item{subset}{A numeric vector with the indexes of a subset of data. Data-points in the subset are heat-mapped and the rest are shown in light grey. By default all data-points are heat-mapped.}

\item{qMap.levels}{The number of levels of the quantile-map (8 by default).}

\item{qMap.cex}{The size of the data-points (as in \code{par()}).}

\item{qMap.bg}{The background colour of the qMap plot. Default value is \code{ptsne.bg = #FFFFFF} (white).}

\item{class.pltt}{If \code{!is.null(labels)} or \code{!is.null(bdm$lbls)}, a colour palette to show class labels with the qMap plots. By default (\code{qMap.pltt = NULL}) the default palette is used.}

\item{qtitle}{A vector of strings with titles for the plots. Default value is \code{qtitle=NULL}.}

\item{cex.main}{The font size of the title (as in \code{par()}).}

\item{colorbar}{A logical value (TRUE by default). FALSE hides the side colorbar.}

\item{layer}{The number of a layer (1 by default).}
}
\value{
None.
}
\description{
Maps quantitative variables onto the embedding space.
}
\details{
This is not a heat-map but a quantile-map plot. This function splits the range of each variable into as many quantiles as specified by \var{levels} so that the color gradient will hardly ever correspond to a constant numeric gradient. Thus, the mapping will show more evenly distributed colors though at the expense of possibly exaggerating artifacts. For variables with very extrem distributions, it will be impossible to find as many quantiles as desired and the distribution of colors will not be so homogeneous.
}
\examples{

bdm.example()
bdm.qMap(ex$map, ex$data)
# --- show only components (1, 2, 4, 8) of the GMM
bdm.qMap(ex$map, ex$data, subset = which(ex$map$lbls \%in\% c(1, 4, 8, 16)))
}
