% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_main.R
\name{bdm.restart}
\alias{bdm.restart}
\title{Restart pt-SNE}
\usage{
bdm.restart(
  data,
  bdm,
  epochs = NULL,
  iters = NULL,
  mpi.cl = NULL,
  threads = NULL,
  layers = NULL,
  info = 0
)
}
\arguments{
\item{data}{Input data (a matrix, a big.matrix or a .csv file name).}

\item{bdm}{A \var{bdm} data mapping instance.}

\item{epochs}{Number of epochs to run. Default value \var{epochs = NULL} runs \var{4 *log(n)} epochs.}

\item{iters}{Number of iters per epoch. Default value \var{iters = NULL} runs \var{4 *log(thread_size)} iters/epoch.}

\item{mpi.cl}{An MPI (inter-node parallelization) cluster as returned by \var{bdm.mpi.start()}. Default value is \var{mpi.cl = NULL}, i.e. a 'SOCK' (intra-node parallelization) cluster is automatically generated.}

\item{threads}{Number of parallel threads (according to data size and hardware resources, i.e. number of cores and available memory). Default value is \var{threads = 4}.}

\item{layers}{Number of layers (\var{minimum} 2, \var{maximum} the number of threads). Default value is \var{layers = 2}.}

\item{info}{Output information: 1 yields inter-round results, 0 disables intermediate results. Default value is 0.}
}
\value{
A \var{bdm} data mapping instance.
}
\description{
Restarts the ptSNE algorithm (runs more epochs).
}
\examples{

# --- load example dataset
bdm.example()
\dontrun{
# --- restart ptSNE
m <- bdm.restart(ex$data, ex$map, epochs = 50)
}
}
