% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdm_main.R
\name{bdm.wtt}
\alias{bdm.wtt}
\title{Watertrack transform (WTT)}
\usage{
bdm.wtt(bdm, layer = 1)
}
\arguments{
\item{bdm}{A \var{bdm} data mapping instance.}

\item{layer}{The ptSNE output layer. Default value is \var{layer = 1}.}
}
\value{
A \var{bdm} data mapping instance.
}
\description{
Starts the WTT algorithm (third setp of the mapping protocol).
}
\details{
This function requires the up-stream step \var{bdm.pakde()}.
}
\examples{

# --- load mapped dataset
bdm.example()
# --- perform WTT
m <- bdm.wtt(ex$map)
# --- plot WTT output
bdm.wtt.plot(m)
}
